fix : Bug in offsets for aux_vmids in realm params
- Updated the realm params structure to take 16 bit
  offsets for aux_vmid instead of 64 bit

+++ test/command/common/command_common_host.c
@@ -109,7 +109,7 @@ uint32_t val_host_realm_create_common(val_host_realm_ts *realm)
     params->flags = realm->flags;
     params->flags1 = realm->flags1;
     for (i = 0; i < realm->num_aux_planes; i++)
-        params->aux_vmid[i] = realm->vmid + i + 1;
+        params->aux_vmid[i] = (uint16_t)(realm->vmid + i + 1);
     for (i = 0; i < realm->num_aux_planes; i++)
         params->aux_rtt_base[i] = realm->rtt_aux_l0_addr[i];
